Fixed gcc warning
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@3858 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ include/gpac/internal/mpd.h
@@ -216,10 +216,9 @@ typedef struct
 {	
 	GF_MPD_COMMON_ATTRIBUTES_ELEMENTS
 
-	/*if not set in MPD, an ID is dynamically generated (u32 cast of the structure)*/
-	u64 id;
-	/*if not set in MPD, an ID is dynamically generated (u32 cast of the structure)*/
-	u64 group;
+	u32 id;
+	/*default value is -1: not set in MPD*/
+	s32 group;
 
 	char *lang;
 	char *content_type;
